requirements.txt
openai
rizaio
anthropic
google-generativeai